move chatUtils to /utils
