+++ src/components/ChatUI.tsx
@@ -31,7 +31,7 @@ import {
   resolvePodAndPort,
   startWorkspacePortForward,
   stopWorkspacePortForward,
-} from './resources/chatUtils';
+} from '../utils/chatUtils';
 
 interface Message {
   id: string;
