+++ src/utils/chatUtils.ts
